requirements.txt
pandas==2.0.3
reportlab==4.0.4
pyinstaller==5.13.0
requests==2.31.0
Pillow==10.0.0
